+++ server/src-lib/Hasura/Server/Migrate.hs
@@ -258,7 +258,8 @@ migrations maybeDefaultSourceConfig dryRun maintenanceMode =
         ++ [| ("3", MigrationPair from3To4 Nothing) |]
         :  migrationsFromFile [5..42]
         ++ [| ("42", MigrationPair from42To43 (Just from43To42)) |]
-        :  migrationsFromFile [44]
+        : migrationsFromFile [44]
+        ++  migrationsFromFile [45]
      )
   where
     runTxOrPrint :: Q.Query -> m ()