+++ server/src-lib/Hasura/Server/SchemaUpdate.hs
@@ -2,45 +2,37 @@
 module Hasura.Server.SchemaUpdate
   ( startSchemaSyncListenerThread
   , startSchemaSyncProcessorThread
-  , EventPayload(..)
-  , SchemaSyncCtx(..)
-  , SchemaSyncEvent(..)
-  , SchemaSyncEventRef
+  , ThreadType(..)
+  , ThreadError(..)
   )
 where
 
 import           Hasura.Logging
 import           Hasura.Metadata.Class
 import           Hasura.Prelude
-import           Hasura.RQL.DDL.Schema       (runCacheRWT)
+import           Hasura.RQL.DDL.Schema         (runCacheRWT)
+import           Hasura.RQL.DDL.Schema.Catalog
 import           Hasura.RQL.Types
 import           Hasura.RQL.Types.Run
-import           Hasura.Server.App           (SchemaCacheRef (..), withSCUpdate)
+import           Hasura.Server.App             (SchemaCacheRef (..), withSCUpdate)
 import           Hasura.Server.Logging
-import           Hasura.Server.Types         (InstanceId (..))
+import           Hasura.Server.Types           (InstanceId (..))
 import           Hasura.Session
 
-import           Control.Monad.Trans.Control (MonadBaseControl)
-import           Control.Monad.Trans.Managed (ManagedT)
+import           Control.Monad.Trans.Control   (MonadBaseControl)
+import           Control.Monad.Trans.Managed   (ManagedT)
 import           Data.Aeson
 import           Data.Aeson.Casing
 import           Data.Aeson.TH
 import           Data.IORef
-#ifndef PROFILING
-import           GHC.AssertNF
-#endif
 
-import qualified Control.Concurrent.Extended as C
-import qualified Control.Concurrent.STM      as STM
-import qualified Control.Immortal            as Immortal
-import qualified Data.Text                   as T
-import qualified Data.Time                   as UTC
-import qualified Database.PG.Query           as PG
-import qualified Database.PostgreSQL.LibPQ   as PQ
-import qualified Network.HTTP.Client         as HTTP
-
-pgChannel :: PG.PGChannel
-pgChannel = "hasura_schema_update"
+import qualified Control.Concurrent.Extended   as C
+import qualified Control.Concurrent.STM        as STM
+import qualified Control.Immortal              as Immortal
+import qualified Data.HashMap.Strict           as HM
+import qualified Data.HashSet                  as HS
+import qualified Database.PG.Query             as Q
+import qualified Network.HTTP.Client           as HTTP
 
 data ThreadType
   = TTListener
@@ -50,7 +42,6 @@ data ThreadType
 instance Show ThreadType where
   show TTListener  = "listener"
   show TTProcessor = "processor"
-
 
 data SchemaSyncThreadLog
   = SchemaSyncThreadLog
@@ -69,23 +60,6 @@ instance ToEngineLog SchemaSyncThreadLog Hasura where
   toEngineLog threadLog =
     (suelLogLevel threadLog, ELTInternal ILTSchemaSyncThread, toJSON threadLog)
 
-data EventPayload
-  = EventPayload
-  { _epInstanceId    :: !InstanceId
-  , _epOccurredAt    :: !UTC.UTCTime
-  , _epInvalidations :: !CacheInvalidations
-  }
-$(deriveJSON hasuraJSON ''EventPayload)
-
-data SchemaSyncEvent
-  = SSEListenStart !UTC.UTCTime
-  | SSEPayload !Value
-
-instance ToJSON SchemaSyncEvent where
-  toJSON = \case
-    SSEListenStart time -> String $ "event listening started at " <> tshow time
-    SSEPayload payload  -> toJSON payload
-
 data ThreadError
   = TEPayloadParse !Text
   | TEQueryError !QErr
@@ -94,17 +68,6 @@ $(deriveToJSON
                  , sumEncoding = TaggedObject "type" "info"
                  }
  ''ThreadError)
-
-type SchemaSyncEventRef = STM.TVar (Maybe SchemaSyncEvent)
-
--- | Context required for schema syncing. Listener thread id,
--- event references and cache init time
-data SchemaSyncCtx
-  = SchemaSyncCtx
-  { _sscListenerThreadId   :: !Immortal.Thread
-  , _sscSyncEventRef       :: !SchemaSyncEventRef
-  , _sscCacheInitStartTime :: !UTC.UTCTime
-  }
 
 logThreadStarted
   :: (MonadIO m)
@@ -124,24 +87,25 @@ logThreadStarted logger instanceId threadType thread =
 When multiple graphql-engine instances are serving on same metadata storage,
 each instance should have schema cache in sync with latest metadata. Somehow
 all instances should communicate each other when any request has modified metadata.
-We make use of Postgres listen-notify to notify metadata updates and listen to those
-events. Each instance is identified by an @'InstanceId'. We choose 'hasura_schema_update'
-as our notify channel. Following steps take place when an API request made to update
-metadata:
 
-1. After handling the request we publish an event via a postgres channel with payload
-   containing instance id and other information to build schema cache.
-   See @'notifySchemaCacheSync'.
+We track the metadata schema version in postgres and poll for this
+value in a thread.  When the schema version has changed, the instance
+will update its local metadata schema and remove any invalidated schema cache data.
 
-2. On start up, before initialising schema cache, an async thread is invoked to
-   continuously listen to Postgres events on 'hasura_schema_update' channel. The payload
-   present in the event is decoded and pushed to a shared @'SchemaSyncEventRef' reference.
-   See @'startSchemaSyncListenerThread'.
+The following steps take place when an API request made to update metadata:
 
-3. Before starting API server, another async thread is invoked to process events pushed
-   by the listener thread via @'SchemaSyncEventRef' reference. Based on the event instance id
-   or listen start time we decide to reload schema cache or not.
-   See @'startSchemaSyncProcessorThread'.
+1. After handling the request we insert the new metadata schema json
+   into a postgres tablealong with a schema version.
+
+2. On start up, before initialising schema cache, an async thread is
+   invoked to continuously poll the Postgres notifications table for
+   the latest metadata schema version. The schema version is pushed to
+   a shared `TMVar`.
+
+3. Before starting API server, another async thread is invoked to
+   process events pushed by the listener thread via the `TMVar`. If
+   the instance's schema version is not current with the freshly
+   updated TMVar version then we update the local metadata.
 
 Why we need two threads if we can capture and reload schema cache in a single thread?
 
@@ -164,20 +128,22 @@ if listen started after schema cache init start time.
 -- See Note [Schema Cache Sync]
 startSchemaSyncListenerThread
   :: C.ForkableMonadIO m
-  => PG.PGPool
-  -> Logger Hasura
+  => Logger Hasura
+  -> Q.PGPool
   -> InstanceId
-  -> ManagedT m (Immortal.Thread, SchemaSyncEventRef)
-startSchemaSyncListenerThread pool logger instanceId = do
-  -- only the latest event is recorded here
-  -- we don't want to store and process all the events, only the latest event
-  schemaSyncEventRef <- liftIO $ STM.newTVarIO Nothing
-
+  -> Milliseconds
+  -> STM.TMVar MetadataResourceVersion
+  -> Bool
+  -- ^ Disable schema listener
+  -> ManagedT m (Immortal.Thread)
+startSchemaSyncListenerThread logger pool instanceId interval metaVersionRef isDisabled = do
   -- Start listener thread
-  listenerThread <- C.forkManagedT "SchemeUpdate.listener" logger . liftIO $
-                    listener pool logger schemaSyncEventRef
+  -- TODO: Move isActive out of schemasync function and eliminate else case
+  listenerThread <- if not isDisabled
+                   then C.forkManagedT "SchemeUpdate.listener" logger $ listener logger pool metaVersionRef interval
+                   else C.forkManagedT "SchemaUpdate.listener" logger . liftIO $ forever $ C.sleep 10
   logThreadStarted logger instanceId TTListener listenerThread
-  pure (listenerThread, schemaSyncEventRef)
+  pure listenerThread
 
 -- | An async thread which processes the schema sync events
 -- See Note [Schema Cache Sync]
@@ -188,62 +154,48 @@ startSchemaSyncProcessorThread
      )
   => Logger Hasura
   -> HTTP.Manager
-  -> SchemaSyncEventRef
+  -> STM.TMVar MetadataResourceVersion
   -> SchemaCacheRef
   -> InstanceId
-  -> UTC.UTCTime
   -> ServerConfigCtx
   -> ManagedT m Immortal.Thread
 startSchemaSyncProcessorThread logger httpMgr
-  schemaSyncEventRef cacheRef instanceId cacheInitStartTime serverConfigCtx = do
+  schemaSyncEventRef cacheRef instanceId serverConfigCtx = do
   -- Start processor thread
   processorThread <- C.forkManagedT "SchemeUpdate.processor" logger $
     processor logger httpMgr schemaSyncEventRef
-              cacheRef instanceId cacheInitStartTime serverConfigCtx
+              cacheRef instanceId serverConfigCtx
   logThreadStarted logger instanceId TTProcessor processorThread
   pure processorThread
 
+-- TODO: This is also defined in multitenant, consider putting it in a library somewhere
+forcePut :: STM.TMVar a -> a -> IO ()
+forcePut v a = STM.atomically $ STM.tryTakeTMVar v >> STM.putTMVar v a
+
+schemaVersionCheckHandler
+  :: Q.PGPool -> STM.TMVar MetadataResourceVersion -> IO (Either QErr ())
+schemaVersionCheckHandler pool metaVersionRef =
+   (runExceptT $
+     Q.runTx pool (Q.RepeatableRead, Nothing) $
+       fetchMetadataResourceVersionFromCatalog) >>= \case
+    Right version ->
+      Right <$> forcePut metaVersionRef version
+    Left err -> pure $ Left err
+
 -- | An IO action that listens to postgres for events and pushes them to a Queue, in a loop forever.
 listener
-  :: PG.PGPool
-  -> Logger Hasura
-  -> SchemaSyncEventRef
-  -> IO void
-listener pool logger updateEventRef =
-  -- Never exits
+  :: MonadIO m
+  => Logger Hasura
+  -> Q.PGPool
+  -> STM.TMVar MetadataResourceVersion
+  -> Milliseconds
+  -> m void
+listener logger pool metaVersionRef interval =
   forever $ do
-    listenResE <-
-      liftIO $ runExceptT $ PG.listen pool pgChannel notifyHandler
-    onLeft listenResE onError
-    logWarn
-    -- Trying to start listening after one second.
-    -- See Note [Schema Cache Sync].
-    C.sleep $ seconds 1
-  where
-    threadType = TTListener
-
-    notifyHandler = \case
-      PG.PNEOnStart        -> do
-        time <- UTC.getCurrentTime
-        STM.atomically $ STM.writeTVar updateEventRef $ Just $ SSEListenStart time
-      PG.PNEPQNotify notif ->
-        case eitherDecodeStrict $ PQ.notifyExtra notif of
-          Left e -> logError logger threadType $ TEPayloadParse $ T.pack e
-          Right payload -> do
-            logInfo logger threadType $ object ["received_event" .= payload]
-#ifndef PROFILING
-            $assertNFHere payload  -- so we don't write thunks to mutable vars
-#endif
-            -- Push a notify event to Queue
-            STM.atomically $ STM.writeTVar updateEventRef $ Just $ SSEPayload payload
-
-    onError = logError logger threadType . TEQueryError
-    -- NOTE: we handle expected error conditions here, while unexpected exceptions will result in
-    -- a restart and log from 'forkImmortal'
-    logWarn = unLogger logger $
-      SchemaSyncThreadLog LevelWarn TTListener $ String
-        "error occurred, retrying postgres listen after 1 second"
-
+    respErr <- liftIO $ schemaVersionCheckHandler pool metaVersionRef
+    liftIO $ do
+      onLeft respErr (logError logger TTListener . TEQueryError)
+      C.sleep (milliseconds interval)
 
 -- | An IO action that processes events from Queue, in a loop forever.
 processor
@@ -254,76 +206,68 @@ processor
      )
   => Logger Hasura
   -> HTTP.Manager
-  -> SchemaSyncEventRef
+  -> STM.TMVar MetadataResourceVersion
   -> SchemaCacheRef
   -> InstanceId
-  -> UTC.UTCTime
   -> ServerConfigCtx
   -> m void
-processor logger httpMgr updateEventRef
-  cacheRef instanceId cacheInitStartTime serverConfigCtx =
-  -- Never exits
-  forever $ do
-    event <- liftIO $ STM.atomically getLatestEvent
-    logInfo logger threadType $ object ["processed_event" .= event]
-    (shouldReload, cacheInvalidations) <- case event of
-      SSEListenStart time ->
-        -- If listening started after cache initialization, just refresh the schema cache unconditionally.
-        -- See Note [Schema Cache Sync]
-        pure $ (time > cacheInitStartTime, mempty)
-      SSEPayload payload -> do
-        eitherResult <- runMetadataStorageT $ processSchemaSyncEventPayload instanceId payload
-        case eitherResult of
-          Left e -> do
-            logError logger threadType $ TEPayloadParse $ qeError e
-            pure (False, mempty)
-          Right SchemaSyncEventProcessResult{..} ->
-            pure (_sseprShouldReload, _sseprCacheInvalidations)
+processor logger httpMgr metaVersionRef
+  cacheRef instanceId serverConfigCtx = forever $ do
+  metaVersion <- liftIO $ STM.atomically $ STM.takeTMVar metaVersionRef
 
-    when shouldReload $
-      refreshSchemaCache logger httpMgr cacheRef cacheInvalidations
-        threadType serverConfigCtx "schema cache reloaded"
-  where
-    -- checks if there is an event
-    -- and replaces it with Nothing
-    getLatestEvent = do
-      eventM <- STM.readTVar updateEventRef
-      case eventM of
-        Just event -> do
-          STM.writeTVar updateEventRef Nothing
-          return event
-        Nothing -> STM.retry
-    threadType = TTProcessor
+  respErr <- runMetadataStorageT $
+    refreshSchemaCache metaVersion instanceId logger httpMgr cacheRef TTProcessor serverConfigCtx
+  onLeft respErr (logError logger TTProcessor . TEQueryError)
 
 refreshSchemaCache
   :: ( MonadIO m
-     , MonadBaseControl IO m
-     , MonadMetadataStorage (MetadataStorageT m)
-     , MonadResolveSource m
-     )
-  => Logger Hasura
+    , MonadBaseControl IO m
+    , MonadMetadataStorage m
+    , MonadResolveSource m
+    )
+  => MetadataResourceVersion
+  -> InstanceId
+  -> Logger Hasura
   -> HTTP.Manager
   -> SchemaCacheRef
-  -> CacheInvalidations
   -> ThreadType
   -> ServerConfigCtx
-  -> Text
   -> m ()
-refreshSchemaCache logger httpManager
-    cacheRef invalidations threadType serverConfigCtx msg = do
-  -- Reload schema cache from catalog
-  eitherMetadata <- runMetadataStorageT fetchMetadata
-  resE <- runExceptT $ do
-    (metadata, _) <- liftEither eitherMetadata
-    withSCUpdate cacheRef logger do
-      rebuildableCache <- fst <$> liftIO (readIORef $ _scrCache cacheRef)
-      ((), cache, _) <- buildSchemaCacheWithOptions CatalogSync invalidations metadata
-        & runCacheRWT rebuildableCache
-        & peelRun runCtx
-      pure ((), cache)
-  case resE of
-    Left e   -> logError logger threadType $ TEQueryError e
-    Right () -> logInfo logger threadType $ object ["message" .= msg]
+refreshSchemaCache resourceVersion instanceId logger httpManager
+    cacheRef threadType serverConfigCtx = do
+      respErr <- runExceptT $ withSCUpdate cacheRef logger $ do
+        rebuildableCache <- fst <$> liftIO (readIORef $ _scrCache cacheRef)
+        (msg, cache, _) <- peelRun runCtx $ runCacheRWT rebuildableCache $ do
+          schemaCache <- askSchemaCache
+          let engineResourceVersion = fromMaybe 0 $ scMetadataResourceVersion schemaCache -- TODO: Can we remove the maybe from scMetadataResourceVersion?
+          unless (engineResourceVersion == resourceVersion) $ do
+            (metadata, latestResourceVersion) <- fetchMetadata
+            notifications <- fetchMetadataNotifications engineResourceVersion instanceId
+
+            logDebug logger threadType $ "DEBUG: refreshSchemaCache Called: engineResourceVersion: " <> show engineResourceVersion <> ", fresh resource version: " <> show latestResourceVersion
+            case notifications of
+              [] -> do
+                logInfo logger threadType $ object ["message" .= ("Schema Version changed with no notifications" :: Text)]
+                setMetadataResourceVersionInSchemaCache latestResourceVersion
+              _ -> do
+                let cacheInvalidations =
+                      if any ((== (engineResourceVersion + 1)) . fst) notifications
+                        then -- If (engineResourceVersion + 1) is in the list of notifications then
+                             -- we know that we haven't missed any.
+                             mconcat $ snd <$> notifications
+                        else -- Otherwise we may have missed some notifications so we need to invalidate the
+                             -- whole cache.
+                             CacheInvalidations
+                               { ciMetadata = True
+                               , ciRemoteSchemas = HS.fromList $ getAllRemoteSchemas schemaCache
+                               , ciSources = HS.fromList $ HM.keys $ scSources schemaCache
+                               }
+                logInfo logger threadType $ object ["currentVersion" .= engineResourceVersion, "latestResourceVersion" .= latestResourceVersion]
+                buildSchemaCacheWithOptions CatalogSync cacheInvalidations metadata
+                setMetadataResourceVersionInSchemaCache latestResourceVersion
+                logInfo logger threadType $ object ["message" .= ("Schema Version changed with notifications" :: Text)]
+        pure (msg, cache)
+      onLeft respErr (logError logger threadType . TEQueryError)
  where
    runCtx = RunCtx adminUserInfo httpManager serverConfigCtx
 
@@ -335,3 +279,7 @@ logError :: (MonadIO m, ToJSON a) => Logger Hasura -> ThreadType -> a -> m ()
 logError logger threadType err =
   unLogger logger $ SchemaSyncThreadLog LevelError threadType $
     object ["error" .= toJSON err]
+
+logDebug :: (MonadIO m) => Logger Hasura -> ThreadType -> String -> m ()
+logDebug logger threadType msg = unLogger logger $
+  SchemaSyncThreadLog LevelDebug threadType $ object ["message" .= msg]